backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
PyJWT==2.8.0
python-multipart==0.0.6
SQLAlchemy-Utils==0.41.1
psycopg2-binary==2.9.9
python-dateutil==2.8.2
requests==2.31.0
email-validator==2.1.0.post1
